# api/openapi-spec/swagger.yaml
openapi: 3.0.0
info:
  title: Audio Converter API
  description: A service that exposes a RESTful API to convert wav to mp3 and vice versa.
  version: 1.0.0
servers:
  - url: http://localhost
    description: Development server
paths:
  /docs:
    get:
      summary: Get documentation
      security:
        - bearerAuth: []
      tags: 
       - Documentation
      responses:
        '200':    
          description: OK
        '401':
          $ref: '#/components/responses/Unauthorized'    
        '500':    
          $ref: '#/components/responses/InternalServerError'
  /user/signup:
    post:
      summary: Create a user
      tags: 
       - Users
      requestBody:
          $ref: '#/components/requestBodies/User'
      responses:
        '201':    
          description: The user has been created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
              example: 
                id: '1fa85f64-5717-4562-b3fc-2c963f66afa5'
        '400':    
          $ref: '#/components/responses/BadRequest' 
        '409':
          description: The user already exists
        '500':    
          $ref: '#/components/responses/InternalServerError'
  /user/login:
    post:
      summary: Authenticate a user
      tags: 
       - Users
      requestBody:
          $ref: '#/components/requestBodies/User'
      responses:
        '200':    
          description: Successful log in
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    format: JWT
              example: 
                token: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.
                eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.
                SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /conversion:
    post:
      summary: Make a request for audio conversion
      security:
        - bearerAuth: []
      tags: 
        - Requests
      requestBody:
          $ref: '#/components/requestBodies/ConversionRequest'
      responses:
        '201':
          description: The request has been created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  request_id:
                    type: string
                    format: uuid
              example: 
                  request_id: '2fa85f64-5717-4562-b3fc-2c963f66afa5'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':   
          $ref: '#/components/responses/InternalServerError'        
  /request_history:
    get:
      summary: Get request history of a user
      security:
        - bearerAuth: []
      tags: 
      - Requests
      responses:
        '200':
          description: Successfully got user's request history
          content: 
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HistoryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /download_audio/{id}:
    get:
      summary: Download the audio by id
      security:
        - bearerAuth: []
      tags:
        - Audios 
      parameters:
        - in: path
          name: id
          required: true
          schema: 
            type: string
            format: uuid
      responses:
        '200':
          description: Downloaded successfully 
          content:
            multipart/form-data:
              schema:
                type: object
                properties:
                  file:
                    type: string
                    format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'     

components:   
  schemas:  
    ConversionRequest:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
        source_id:
          type: string
          format: uuid
        source_format:
          $ref: '#/components/schemas/Format'
        target_id:
          type: string
          format: uuid  
        target_format:
          $ref: '#/components/schemas/Format'
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/Status'
    HistoryResponse:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
        audio_name:
          type: string
        source_format:
          $ref: '#/components/schemas/Format'
        target_format:
          $ref: '#/components/schemas/Format'
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/Status'
      example:
          request_id: '3fa85f64-5717-4562-b3fc-2c963f66afa5'
          audio_name: 'Euphoria.wav'
          source_format: wav
          target_format: mp3
          created: 2020-02-20 T 11:32:28 Z
          updated: 2020-02-20 T 13:27:03 Z 
          status: done
    Format: 
        type: string
        enum: [mp3, wav]
    Status:
        type: string
        enum: [queued, processing, done]
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message

  securitySchemes:
    bearerAuth: 
      type: http
      scheme: bearer
      bearerFormat: JWT 
        
  requestBodies:
    User:
      description: A JSON object containing user's credentials
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              username:
                type: string
              password:
                type: string
                format: password
          example:
            username: AudioUser
            password: qwerty1234
    ConversionRequest:
      description: A JSON object containing file and target format
      required: true
      content:
        multipart/form-data:
          schema:
            type: object
            properties:
              file:
                type: string
                format: binary
              source_format:
                $ref: '#/components/schemas/Format'
              target_format:
                $ref: '#/components/schemas/Format'
          example:
            file: some binary sequence
            source_format: mp3
            target_format: wav
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'       
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'    
    InternalServerError:
      description: The server encountered an unexpected condition which prevented it from fulfilling the request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'  